Databaser/requirements.txt
pytest
redis
mysql.connector